requirements.txt
Django==5.2.4
djangorestframework==3.16.0
plotly==6.0.0
tensorflow-cpu==2.20.0
joblib==1.4.2
scikit-learn==1.6.1
pandas==2.3.2
numpy==2.3.3
matplotlib==3.10.1
gunicorn==23.0.0
whitenoise==6.9.0
openpyxl==3.1.5
